Ensure LF line endings for .sh and .R files
If a Windows user has autocrlf=true then when they
checkout files, LF will be converted to CRLF which
messes up the Cygwin scripts.

This commit adds a .gitattributes file which uses
the autocrlf behavior for text files, but not for
.sh and .R files.

It also makes sure we don't interpret .RData as
text. It's binary.